moved SelfISsuedTokenVerifier to identity-hub-core
